+++ .env
@@ -1,0 +1,1 @@
+export APP_SETTINGS="testing"
