+++ tests/test_auth.py
@@ -1,0 +1,4 @@
+import unittest
+import json
+from api.models import User, Question, Answer
+from api import app
